requirements.txt
numpy
pandas
matplotlib
scikit-learn
bokeh
plotly
scipy
pillow
streamlit
torch
diffusers
transformers
fastapi